# docs/swagger.yaml
definitions:
...
  disability_system_backend_internal_modules_notificaciones_dto.CrearNotificacionRequest:
    properties:
      id_incapacidad:
        type: integer
      id_usuario:
        type: integer
      mensaje:
        type: string
      tipo_notificacion:
        type: string
    required:
    - id_usuario
    - mensaje
    - tipo_notificacion
    type: object
...
info:
  contact: {}
paths:
...
  /notificaciones:
    get:
      consumes:
      - application/json
      description: Lista las notificaciones del usuario
      parameters:
      - description: Filtrar por usuario
        in: query
        name: id_usuario
        type: integer
      - description: Filtrar por incapacidad
        in: query
        name: id_incapacidad
        type: integer
      - description: Filtrar por tipo
        in: query
        name: tipo_notificacion
        type: string
      - description: Filtrar por estado de lectura
        in: query
        name: leida
        type: boolean
      - default: 1
        description: Página
        in: query
        name: page
        type: integer
      - default: 20
        description: Límite
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar notificaciones
      tags:
      - notificaciones
    post:
      consumes:
      - application/json
      description: Crea una nueva notificación
      parameters:
      - description: Datos de la notificación
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disability_system_backend_internal_modules_notificaciones_dto.CrearNotificacionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear notificación
      tags:
      - notificaciones
  /notificaciones/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina una notificación
      parameters:
      - description: ID de la notificación
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar notificación
      tags:
      - notificaciones
    get:
      consumes:
      - application/json
      description: Obtiene los detalles de una notificación
      parameters:
      - description: ID de la notificación
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener notificación por ID
      tags:
      - notificaciones
  /notificaciones/{id}/leida:
    patch:
      consumes:
      - application/json
      description: Marca una notificación específica como leída
      parameters:
      - description: ID de la notificación
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Marcar notificación como leída
      tags:
      - notificaciones
  /notificaciones/marcar-todas-leidas:
    patch:
      consumes:
      - application/json
      description: Marca todas las notificaciones como leídas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Marcar todas como leídas
      tags:
      - notificaciones
  /notificaciones/no-leidas/count:
    get:
      consumes:
      - application/json
      description: Obtiene el conteo de notificaciones no leídas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Contar notificaciones no leídas
      tags:
      - notificaciones
...
swagger: "2.0"